[Moore][ImportVerilog] Introduce new operations: PortOp & NetOp into Moore dialect
PortOp provides functions for port declarations. NetOp is for variable declaration. In the original design,
both portOp and netOp (variableOp) are needed to represent a port. The generated Moore IR would be redundant
and lead to the poor translation consistency between non-ANSI and ANSI declaration. We split the original
logic of representing ports in port list. The PortOp can now idependently represent information such as type,
direction, and named SSA value which is used to interact with instances or other operations in a Moore module.
Several ANSI and non-ANSI declaration examplels were provided in the test file (basic.sv in ImportVerilog directory).

Co-authored-by: Fabian Schuiki <fabian@schuiki.ch>
Co-authored-by: Hailong Sun <hailong.sun@terapines.com>